Add animation loading

+++ frontend/src/app/guards/auth.guard.ts
@@ -18,27 +18,31 @@ export class AuthGuard implements CanActivate {
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot
   ): boolean {
-      const expectedRole = next.data['expectedRole'];
+    const expectedRole = next.data['expectedRole'];
 
-      const token = this.authService.decodedToken()
+    if (!this.authService.isSignIned()) {
+      Swal.fire({
+        title: 'Access Denied',
+        text: 'Please sign in before access',
+        icon: 'warning',
+      });
+      this.router.navigate(['signin']);
+      return false;
+    }
 
-      if (expectedRole && token.role !== expectedRole) {
-        Swal.fire({
-          title: 'Access Denied',
-          text: 'Please sign in with correct role before access',
-          icon: 'warning',
-        });
-        this.router.navigate(['signin']);
-        localStorage.clear()
-        return false;
-      }
+    const token = this.authService.decodedToken();
+
+    if (expectedRole && token.role !== expectedRole) {
+      Swal.fire({
+        title: 'Access Denied',
+        text: 'Please sign in with correct role before access',
+        icon: 'warning',
+      });
+      this.router.navigate(['signin']);
+      localStorage.clear();
+      return false;
+    }
 
     return true;
   }
 }
-
-// if (!this.authService.isSignIned()) {
-//   this.authService.signOut();
-//   this.router.navigate(['signin']);
-//   return false;
-// }
